+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Changelog for insanic
 0.5.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- CHORE: refactor content-length headers for responses.
+- CHORE: rename json to json_response
 
 
 0.5.4 (2018-07-17)